Prescription Creation with more fields!
