+++ server/service/src/invoice/prescription/insert/generate.rs
@@ -5,7 +5,7 @@ use repository::{
     RepositoryError, StorageConnection,
 };
 
-use crate::number::next_number;
+use crate::{invoice::invoice_date_utils::handle_new_backdated_datetime, number::next_number};
 
 use super::InsertPrescription;
 
@@ -19,6 +19,8 @@ pub fn generate(
         diagnosis_id,
         program_id,
         their_reference,
+        clinician_id,
+        prescription_date,
     }: InsertPrescription,
 ) -> Result<InvoiceRow, RepositoryError> {
     let current_datetime = Utc::now().naive_utc();
@@ -27,7 +29,7 @@ pub fn generate(
         .pop()
         .ok_or(RepositoryError::NotFound)?;
 
-    let result = InvoiceRow {
+    let mut invoice = InvoiceRow {
         id,
         user_id: Some(user_id.to_string()),
         name_link_id: patient_id,
@@ -54,7 +56,7 @@ pub fn generate(
         cancelled_datetime: None,
         linked_invoice_id: None,
         requisition_id: None,
-        clinician_link_id: None,
+        clinician_link_id: clinician_id,
         original_shipment_id: None,
         backdated_datetime: None,
         diagnosis_id,
@@ -65,5 +67,9 @@ pub fn generate(
         is_cancellation: false,
     };
 
-    Ok(result)
+    if let Some(date) = prescription_date {
+        handle_new_backdated_datetime(&mut invoice, date, current_datetime);
+    }
+
+    Ok(invoice)
 }
