+++ server/service/src/invoice/prescription/insert/mod.rs
@@ -2,6 +2,7 @@ use crate::activity_log::activity_log_entry;
 use crate::invoice::query::get_invoice;
 use crate::service_provider::ServiceContext;
 use crate::WithDBError;
+use chrono::NaiveDateTime;
 use repository::{ActivityLogType, Invoice};
 use repository::{InvoiceRowRepository, RepositoryError};
 
@@ -17,6 +18,8 @@ pub struct InsertPrescription {
     pub diagnosis_id: Option<String>,
     pub program_id: Option<String>,
     pub their_reference: Option<String>,
+    pub clinician_id: Option<String>,
+    pub prescription_date: Option<NaiveDateTime>,
 }
 
 #[derive(Debug, PartialEq)]
