+++ server/service/src/invoice/prescription/update/generate.rs
@@ -1,13 +1,16 @@
-use chrono::{NaiveDateTime, Utc};
+use chrono::Utc;
 
 use repository::{
     EqualFilter, InvoiceLineFilter, InvoiceLineRepository, InvoiceLineRow, InvoiceRow,
     InvoiceStatus, RepositoryError, StockLineRow, StorageConnection,
 };
 
-use crate::invoice::common::{
-    generate_batches_total_number_of_packs_update, get_invoice_status_datetime,
-    InvoiceLineHasNoStockLine,
+use crate::invoice::{
+    common::{
+        generate_batches_total_number_of_packs_update, get_invoice_status_datetime,
+        InvoiceLineHasNoStockLine,
+    },
+    invoice_date_utils::handle_new_backdated_datetime,
 };
 
 use super::{UpdatePrescription, UpdatePrescriptionError, UpdatePrescriptionStatus};
@@ -51,7 +54,6 @@ pub(crate) fn generate(
     set_new_status_datetime(&mut update_invoice, &input_status);
 
     update_invoice.name_link_id = input_patient_id.unwrap_or(update_invoice.name_link_id);
-    // update_invoice.clinician_link_id = input_clinician_id.or(update_invoice.clinician_link_id);
     if let Some(clinician_link_id) = input_clinician_id {
         update_invoice.clinician_link_id = clinician_link_id.value;
     }
@@ -111,29 +113,6 @@ fn should_update_batches_total_number_of_packs(
             && invoice_status_index < InvoiceStatus::Picked.index()
     } else {
         false
-    }
-}
-// Replace datetimes that are not null with the new status_datetime
-fn replace_status_datetimes(invoice: &mut InvoiceRow, new_status_datetime: NaiveDateTime) {
-    invoice.allocated_datetime = invoice.allocated_datetime.map(|_| new_status_datetime);
-    invoice.picked_datetime = invoice.picked_datetime.map(|_| new_status_datetime);
-    invoice.verified_datetime = invoice.verified_datetime.map(|_| new_status_datetime);
-}
-
-// Handle a change to backdated_time
-fn handle_new_backdated_datetime(
-    invoice: &mut InvoiceRow,
-    backdated_datetime: NaiveDateTime,
-    now: NaiveDateTime,
-) {
-    if backdated_datetime > now {
-        // If the backdated_datetime is in the future, we unset the backdated_datetime as it isn't possible to future date.
-        invoice.backdated_datetime = None;
-        replace_status_datetimes(invoice, now);
-    } else {
-        // Otherwise, we need to update the backdated_datetime to the new one, and replace existing status times
-        invoice.backdated_datetime = Some(backdated_datetime);
-        replace_status_datetimes(invoice, backdated_datetime);
     }
 }
 
@@ -206,78 +185,4 @@ fn lines_to_trim(
         .map(|l| l.invoice_line_row)
         .collect();
     Ok(Some(invoice_line_rows))
-}
-
-#[cfg(test)]
-mod test {
-    use chrono::Utc;
-    use repository::InvoiceRow;
-    use repository::InvoiceStatus;
-    use repository::InvoiceType;
-
-    #[actix_rt::test]
-    async fn handle_new_backdated_datetime_test() {
-        let now = Utc::now().naive_utc();
-        // Create a new invoice 2 days ago
-        let invoice_time = Utc::now().naive_utc() - chrono::Duration::days(2);
-
-        let mut invoice = InvoiceRow {
-            id: "test_invoice_id".to_string(),
-            status: InvoiceStatus::Picked,
-            created_datetime: invoice_time,
-            allocated_datetime: Some(invoice_time),
-            picked_datetime: Some(invoice_time),
-            verified_datetime: None,
-            backdated_datetime: None,
-            name_link_id: "test_patient_id".to_string(),
-            clinician_link_id: None,
-            comment: None,
-            colour: None,
-            name_store_id: None,
-            store_id: String::new(),
-            user_id: None,
-            invoice_number: 0,
-            r#type: InvoiceType::Prescription,
-            on_hold: false,
-            their_reference: None,
-            transport_reference: None,
-            shipped_datetime: None,
-            delivered_datetime: None,
-            requisition_id: None,
-            linked_invoice_id: None,
-            tax_percentage: None,
-            currency_id: None,
-            currency_rate: 0.0,
-            original_shipment_id: None,
-            ..Default::default()
-        };
-
-        // Check that we can backdate to 3 days ago
-        let backdated_datetime = Utc::now().naive_utc() - chrono::Duration::days(3);
-        super::handle_new_backdated_datetime(&mut invoice, backdated_datetime, now);
-
-        assert_eq!(invoice.backdated_datetime, Some(backdated_datetime));
-        assert_eq!(invoice.allocated_datetime, Some(backdated_datetime));
-        assert_eq!(invoice.picked_datetime, Some(backdated_datetime));
-        assert_eq!(invoice.verified_datetime, None);
-
-        // Check that we can't backdate to tomorrow, this should unset the backdated_datetime
-        // and set the status times to now
-        let backdated_datetime = Utc::now().naive_utc() + chrono::Duration::days(1);
-        super::handle_new_backdated_datetime(&mut invoice, backdated_datetime, now);
-
-        assert_eq!(invoice.backdated_datetime, None);
-        assert_eq!(invoice.allocated_datetime, Some(now));
-        assert_eq!(invoice.picked_datetime, Some(now));
-        assert_eq!(invoice.verified_datetime, None);
-
-        // Check that we can backdate to 2 days ago
-        let backdated_datetime = Utc::now().naive_utc() - chrono::Duration::days(2);
-        super::handle_new_backdated_datetime(&mut invoice, backdated_datetime, now);
-
-        assert_eq!(invoice.backdated_datetime, Some(backdated_datetime));
-        assert_eq!(invoice.allocated_datetime, Some(backdated_datetime));
-        assert_eq!(invoice.picked_datetime, Some(backdated_datetime));
-        assert_eq!(invoice.verified_datetime, None);
-    }
 }
